#49 - Replaced "MODIFIER LETTER SMALL TOP HALF O" with "DEGREE SIGN".
a) There is no way to prevent a dead-key character from emitting a code point after a non matching combination in the dead-key table.
b) A solution is to attempt to find all possible combinations, with all available code points from the base layer and map them to a NUL character.
     This effectively deactivates said combination. However, BACKSPACE, TAB and DELETE keys cannot be mapped, and will leak the code point.

Therefore, for simplicity, I do not bother deactivated every anticipated combinations.

However, I choose a glyph that is not too different that the intended "INVERTED BREVE" character in the shape of the "DEGREE SIGN".
The approximate character was "MODIFIER LETTER SMALL TOP HALF O" (U+1D54) but kbdutool.exe automatically translates it into a ligature.
Therefore, the dead-key no longer worked for this character. The "DEGREE SIGN" is safe and restore the intended behavior.