Replace "LATIN SMALL LETTER SCHWA" with "LATIN SMALL LETTER LONG S".
Since there is not way to prevent a non-matching dead key character to emit a code point,
We have to choose a letter. In the European Extended Characters layer, the "LATIN SMALL LETTER LONG S"
can also be combined by a diacritical mark. Therefore, when choosing the diacritical mark (SHORT SOLIDUS OVERLAY or DOT ABOVE),
the corresponding precomposed letter is emitted.